Refactor
Move routes into different module
Some fixes and renames
