+++ docs/api/bans/post.md
@@ -16,7 +16,7 @@ Bans target with provided details.
 
 | Field    | Type                                  | Is required | Note               |
 |----------|---------------------------------------|-------------|--------------------|
-| `target` | `{ ip: string?, user-agent:string? }` | Yes         | Ban target         | 
+| `target` | `{ ip: string?, user_agent:string? }` | Yes         | Ban target         | 
 | `reason` | `string`                              | Yes         | Ban reason         |
 | `ttl`    | `number`                              | Yes         | Ban TTL, in seconds |
 
@@ -26,7 +26,7 @@ Bans target with provided details.
 {
     "target": {
         "ip": "11.12.13.14",
-        "user-agent": "curl user-agent"
+        "user_agent": "curl user-agent"
     },
     "reason": "Exceeded requests per minute limit, excess 2.0",
     "ttl": 300
