+++ src/api/routes/mod.rs
@@ -1,0 +1,7 @@
+pub use admin::admin_settings;
+pub use ban::process_ban;
+pub use unban::process_unban;
+
+mod admin;
+mod ban;
+mod unban;
