+++ src/config.rs
@@ -1,7 +1,7 @@
 use pepe_config::{ConfigError, FileFormat};
 use serde::{Deserialize, Serialize};
 
-use crate::server;
+use crate::api;
 use crate::telemetry;
 
 pub const DEFAULT_CONFIG: &str = include_str!("../config.yaml");
@@ -10,7 +10,7 @@ pub const DEFAULT_CONFIG: &str = include_str!("../config.yaml");
 pub struct Config {
     pub dry_run: Option<bool>,
     pub redis: pepe_config::redis::Config,
-    pub server: server::Config,
+    pub server: api::Config,
     pub telemetry: telemetry::Config,
     pub redis_keys_prefix: String,
     #[serde(default = "default_redis_query_timeout_secs")]
