+++ src/lib.rs
@@ -1,3 +1,4 @@
+pub mod api;
 pub mod ban_hammer;
 pub mod config;
 pub mod dry_runner;
@@ -5,7 +6,4 @@ pub mod errors;
 pub mod http_error;
 pub mod model;
 pub mod redis;
-pub mod server;
 pub mod telemetry;
-
-const ANALYZER_HEADER: &str = "X-Analyzer-Id";
