--- src/server/mod.rs
@@ -1,6 +1,0 @@
-pub mod config;
-#[allow(clippy::module_inception)]
-pub mod server;
-
-pub use config::Config;
-pub use server::Server;
